unstable update for tonight: grid
the beginning of the gridpocalypse. new idea is to do something tetris-like so that sand particles (which actually have a width and height!) don't overlap, the jank way. hoping this grid issue will also overcome lag in the future, as i will be able to store the color values of non-moving sand particles in the grid so that way i don't have 230948 objects floating around when the user is trying to have a fun time